Rework testable code into importable modules
